Removing duplicate DeferCondition type
